refactor: renamed component
